mvtec-ad/requirements.txt
numpy==1.21.2
Pillow==9.5.0
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.6.2
torch==1.13.1
torchvision==0.6.1
tqdm==4.65.0
visdom==0.2.1
